requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
oauthlib>=3.2.2
mega.py>=1.0.8
requests>=2.31.0
pycryptodome>=3.19.0
google-api-python-client>=2.119.0
google-auth>=2.27.0
google-auth-httplib2>=0.2.0
